requirements.txt
# Existing packages
asgiref==3.7.2
Django==5.0
django-admin-interface==0.26.1
django-colorfield==0.9.0
django-cors-headers==4.3.1
django-filter==23.5
dj-rest-auth==5.0.1
django-rest-authtoken==2.1.2
djangorestframework==3.15.1
Markdown==3.5.2
pymysql==1.1.1
Pillow==10.4.0
python-slugify==8.0.4
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
#tzdata==2024.1
tzdata

# Add these to your existing requirements.txt
channels==4.0.0
channels-redis==4.2.0
redis==5.0.1
celery==5.5.3
django-celery-beat==2.8.1
django-celery-results==2.6.0
django-redis==5.4.0

# New packages for wallet system
stripe==7.8.0
django-money==3.4.1
py-moneyed==3.0
